Created some integration testcases for panel space
