+++ pytest.ini
@@ -2,3 +2,5 @@
 pythonpath = .
 testpaths = test_vimmo
 python_files = test_*.py
+markers =
+    integration: integration tests that call external services
