--- vimmo.log
@@ -1,15 +1,0 @@
-2024-12-20 12:08:08,964 - VIMMO - INFO - Starting VIMMO
-2024-12-20 12:08:08,964 - VIMMO - DEBUG - Attempting to connect to the database.
-2024-12-20 12:08:08,964 - VIMMO - INFO - Shutting down VIMMO
-2024-12-20 12:16:42,621 - VIMMO - INFO - Everything imported
-2024-12-20 12:16:43,216 - VIMMO - INFO - Everything imported
-2024-12-20 12:17:54,883 - VIMMO - INFO - Everything imported
-2024-12-20 12:18:41,740 - VIMMO - INFO - Everything imported
-2024-12-20 12:18:42,264 - VIMMO - INFO - Everything imported
-2024-12-20 12:19:03,654 - VIMMO - INFO - Starting VIMMO
-2024-12-20 12:19:03,654 - VIMMO - INFO - Shutting down VIMMO
-2024-12-20 12:20:14,642 - VIMMO - DEBUG - Everything imported
-2024-12-20 12:20:15,154 - VIMMO - DEBUG - Everything imported
-2024-12-20 12:21:30,880 - VIMMO - DEBUG - Everything imported
-2024-12-20 12:21:37,494 - VIMMO - DEBUG - Everything imported
-2024-12-20 12:23:26,152 - VIMMO - DEBUG - Everything imported
